Added support for `homeKitFanType` field in config.json
There are currently two HomeKit fan types: `Fan` and `Fanv2`. The `Fanv2` type will only work on iOS 11 and if used, the fan won't appear in the accessory list in the Home app on iOS 10. For maximum compatibility use the `Fan` type in the `homeKitFanType` config field. This is also currently the default value for this field if it's not specified.